bond-bot/requirements.txt
aiogram==3.13.1
fastapi==0.115.2
uvicorn==0.30.6
SQLAlchemy==2.0.35
pydantic==2.9.2
python-dotenv==1.0.1
alembic==1.13.3
Jinja2==3.1.4
itsdangerous==2.2.0
Pillow==10.4.0
openpyxl==3.1.5
python-multipart==0.0.12
aiosqlite==0.20.0
httpx==0.27.2
